auvsi/client/requirements.txt
futures
python-dateutil
requests
